continued - need to test getShipTracks and all be fine and done (todo - messenger things) and done all feature

+++ src/main/java/pl/bartlomiej/marineunitmonitoring/point/ActivePointsManager.java
@@ -2,14 +2,13 @@ package pl.bartlomiej.marineunitmonitoring.point;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import pl.bartlomiej.marineunitmonitoring.ais.AisService;
 import pl.bartlomiej.marineunitmonitoring.common.error.MmsiConflictException;
 import pl.bartlomiej.marineunitmonitoring.common.error.NotFoundException;
 import pl.bartlomiej.marineunitmonitoring.shiptracking.ShipTrackHistoryService;
 import pl.bartlomiej.marineunitmonitoring.user.nested.trackedship.TrackedShipService;
 
+import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -22,10 +21,8 @@ import static pl.bartlomiej.marineunitmonitoring.point.PointServiceImpl.UNKNOWN_
 @Component
 public class ActivePointsManager {
     private static final Set<ActivePoint> activePoints = new HashSet<>();
-    private static final int CLEAR_INACTIVE_SHIPS_DELAY = 1000 * 60 * 45;
     private final TrackedShipService trackedShipService;
     private final ShipTrackHistoryService shipTrackHistoryService;
-    private final AisService aisService;
 
     public static List<Long> getMmsis() {
         return of(activePoints.stream()
@@ -50,14 +47,6 @@ public class ActivePointsManager {
                 .isEmpty();
     }
 
-    public static void addActivePoint(ActivePoint activePoint) {
-        if (!isPointActive(activePoint.mmsi)) {
-            activePoints.add(activePoint);
-        } else {
-            log.info("Point is already in list.");
-        }
-    }
-
     public static void removeActivePoint(Long mmsi) {
         if (!activePoints.removeIf(activePoint ->
                 activePoint.mmsi().equals(mmsi))) {
@@ -65,26 +54,35 @@ public class ActivePointsManager {
         }
     }
 
-    @Scheduled(initialDelay = 0, fixedDelay = CLEAR_INACTIVE_SHIPS_DELAY)
-    private void clearInactiveShips() {
-        aisService.fetchLatestShips()
-                .map(aisShip -> aisShip.properties().mmsi())
-                .collectList()
-                .doOnError(error -> log.error("Something go wrong on clearing inactive ships: {}", error.getMessage()))
-                .subscribe(list -> {
+    public void addActivePoint(ActivePoint activePoint) {
+        if (!isPointActive(activePoint.mmsi)) {
+            activePoints.add(activePoint);
+        } else {
+            log.info("Point is already in list.");
+        }
+    }
 
-                    // identify inactive mmsis
-                    List<Long> invalidMmsis = getMmsis();
-                    invalidMmsis.removeAll(list);
+    public void filterInactiveShips(List<Long> activeMmsis) {
+        List<Long> actualMmsis = new ArrayList<>(getMmsis());
+        if (!actualMmsis.isEmpty()) {
+            List<Long> inactiveMmsis = activeMmsis.stream()
+                    .filter(activeMmsi -> !actualMmsis.contains(activeMmsi))
+                    .toList();
 
-                    // remove inactive ships from app
-                    invalidMmsis.forEach(mmsi -> {
-                        removeActivePoint(mmsi);
-                        trackedShipService.removeTrackedShip(mmsi);
-                        shipTrackHistoryService.clearShipHistory(mmsi).subscribe();
-                        log.info("Inactive ships has been removed - {}", mmsi);
-                    });
-                });
+            if (inactiveMmsis.isEmpty()) {
+                log.info("All points are active.");
+            } else {
+                inactiveMmsis
+                        .forEach(mmsi -> {
+                            log.info("Removing inactive point - {}", mmsi);
+                            removeActivePoint(mmsi);
+                            trackedShipService.removeTrackedShip(mmsi);
+                            shipTrackHistoryService.clearShipHistory(mmsi).subscribe();
+                        });
+            }
+        } else {
+            log.info("No points to filter.");
+        }
     }
 
     public record ActivePoint(Long mmsi, String name) {
